More on 3D visualization, some re-factoring
