+++ emhub/data/processing/base.py
@@ -75,6 +75,9 @@ class SessionData:
     def relpath(self, path):
         return os.path.relpath(path, self._path)
 
+    def exists(self, path):
+        return os.path.exists(self.join(path))
+
     def getEpuData(self):
         if not self._epuData:
             epuFolder = self.join('EPU')
